Add subscription alert and default message customization
Updated the TwitchAlertSettings and the TwitchAlertListener to allow for easier message customization. In TwitchAlertSettings, default values for the optional messages were added to avoid potential null reference issues. For the TwitchAlertListener, added logic for formatting alert messages with user's display name before sending the chat message. This change allows for a more personalized alert message.

Also, created a new file for Command `subalert` to set, clear, or reset the alert message. In addition, a reference to FloppyBot.Base.Extensions have been added to both projects `FloppyBot.Aux.TwitchAlerts.Core.csproj` and `FloppyBot.Commands.Aux.Twitch.csproj` to make use of the newly created ObjectExtensions for a more efficient retrieval of properties.

This overall modification provides streamers with the ability to customize the alert message when someone subscribes, enhancing user interaction and experience.