Add integration tests for FloppyBot project
New separate integration test project called FloppyBot.IntTest has been added to the solution. The test project contains tests for MongoDB and Redis functionalities. It utilizes Testcontainers for spinning up MongoDB and Redis in Docker from within the fixture of xUnit tests. This set up ensures a clean and independent data environment for each test run. This implementation will help in maintaining and enhancing the quality of the software.